+++ src/werkzeug/datastructures/range.py
@@ -268,7 +268,8 @@ class ContentRange:
         units: str | None = "bytes",
     ) -> None:
         """Simple method to update the ranges."""
-        assert is_byte_range_valid(start, stop, length), "Bad range provided"
+        if not is_byte_range_valid(start, stop, length):
+            raise ValueError("Bad range provided")
         self._units: str | None = units
         self._start: int | None = start
         self._stop: int | None = stop